wyrzucenie przeliczania indeksow, 2opt powinien dzialac

+++ Optimization/OptimizationParameters.cs
@@ -30,7 +30,7 @@ namespace Optimization
         public string CrossoverMethod { get; set; }
         public string EliminationMethod { get; set; }
         public bool CanMutate { get; set; }
-        public int MutationProbability { get; set; }
+        public double MutationProbability { get; set; }
         public int PopulationSize { get; set; }
         public int ChildrenPerGeneration { get; set; }
         public int TerminationValue { get; set; }
